retrace: Move stuff from retrace to util
The goal of this commit is to free retrace.py from stuff that
can be placed elsewhere (util.py) and make it more structured.

Most of the moved functions in this commit are never used in retrace.py.
