+++ src/index.wsgi
@@ -2,10 +2,9 @@
 from webob import Request
 
 from retrace.retrace import (get_active_tasks,
-                             get_supported_releases,
-                             parse_http_gettext,
-                             response)
+                             get_supported_releases)
 from retrace.config import Config
+from retrace.util import parse_http_gettext, response
 
 CONFIG = Config()
