+++ src/retrace-server-reposync
@@ -15,15 +15,14 @@ from functools import cmp_to_key
 from subprocess import run, DEVNULL, PIPE
 
 from retrace.retrace import (get_canon_arch,
-                             lock,
                              log_debug,
                              log_error,
                              log_info,
-                             log_warn,
-                             parse_rpm_name,
-                             unlock)
+                             log_warn)
+
 from retrace.config import Config
 from retrace.plugins import Plugins
+from retrace.util import lock, unlock, parse_rpm_name
 
 sys.path.insert(0, "/usr/share/retrace-server")
