+++ src/retrace/__init__.py
@@ -1,7 +1,8 @@
-__all__ = ["argparser", "config", "plugins", "retrace", "retrace_worker"]
+__all__ = ["argparser", "config", "plugins", "retrace", "retrace_worker", "util"]
 
 from . import argparser
 from . import config
 from . import plugins
 from . import retrace
 from . import retrace_worker
+from . import util
